Remove some redundant code
This commit makes the SubparsersImpl class extend SubparserGroupImpl.
This makes sense, since a lot of functionality is shared among them and
now only needs to be implemented within SubparserGroupImpl.
